Removing PluMAMain.{cxx,hxx} and replacing in PluMA.{cxx,hxx}. Adding dependency download fucntion to R language. Adding searching for Perl requirements.
