+++ .editorconfig
@@ -17,3 +17,6 @@ indent_size = 4
 
 [.txt]
 insert_final_newline = false
+
+[Dockerfile]
+indent_size = 2
