+++ SConstruct
@@ -150,7 +150,7 @@ env = Environment(
     CXX=getenv("CXX", "c++"),
     CPPDEFINES=["HAVE_PYTHON"],
     SHCCFLAGS=["-fpermissive", "-fPIC", "-I.", "-O2"],
-    SHCXXFLAGS=["-std=c++11", "-fPIC", "-I.", "-O2"], # Should update all plugins to C++17...
+    SHCXXFLAGS=["-std=c++17", "-fPIC", "-I.", "-O2"], # Should update all plugins to C++17...
     CCFLAGS=["-fpermissive", "-fPIC", "-I.", "-O2"],
     CXXFLAGS=["-std=c++17", "-fPIC", "-O2"],
     CPPPATH=include_search_path,
